Minor fix in scanner example
