+++ picas/util.py
@@ -2,7 +2,6 @@
 @author Joris Borgdorff
 """
 
-import argparse
 import time
 from copy import deepcopy
 
@@ -28,18 +27,6 @@ def merge_dicts(dict1, dict2):
 def seconds():
     """get time in seconds"""
     return int(time.time())
-
-
-def arg_parser():
-    """
-    Arguments parser for optional values of the example
-    returns: argparse object
-    """
-    parser = argparse.ArgumentParser(description="Arguments used in the different classes in the example.")
-    parser.add_argument("--design_doc", default="Monitor", type=str, help="Select the designdoc used by the actor class")
-    parser.add_argument("--view", default="todo", type=str, help="Select the view used by the actor class")
-    parser.add_argument("-v", "--verbose", action="store_true", help="Set verbose")
-    return parser
 
 
 class Timer:
